// apps/customer/src/modules/customer/enums/errorCode.enum.ts
/**
 * Profile Error Codes
 */
export enum CUSTOMER_ERROR_CODE {
  PROFILE_REGISTRATION_FAILED = 1101,
  PROFILE_LOGIN_FAILED = 1102,
  PROFILE_UPDATE_FAILED = 1103,
  PROFILE_GET_FAILED = 1104,
  PROFILE_QUERY_FAILED = 1105,
  PROFILE_REGISTRATION_EXISTING_EMAIL = 1106,
  PROFILE_REGISTRATION_STORE_ADDED = 1107,
  PROFILE_LOGIN_NO_STORE_PERMISSION = 1108,
  PROFILE_REGISTRATION_TENANT_ERR = 1109,
  CHILD_ADDITION_FAILED = 1110,
  PROFILE_NOT_FOUND = 1111,
  EMAIL_ALREADY_EXISTS = 1112,
  EMAIL_UPDATE_SUCCESSFUL_CART_UPDATE_FAILED = 1113,
  EMAIL_CHANGE_FAILED = 1114,
}

/**
 * Profile Error Case
 */
export enum CUSTOMER_ERROR_CASE {
  CHANNEL_NOT_FOUND = 'CHANNEL_NOT_FOUND',
  CHANNEL_NOT_PASSED = 'CHANNEL_NOT_PASSED',
  PROFILE_REGISTRATION_STORE_ADDED = 'PROFILE_REGISTRATION_STORE_ADDED',
  PROFILE_REGISTRATION_EXISTING_EMAIL = 'PROFILE_REGISTRATION_EXISTING_EMAIL',
  MASTER_INACTIVE_CUSTOMER = 'MASTER_INACTIVE_CUSTOMER',
  MASTER_CANT_ADD_PARENT_AS_CHILD = 'MASTER_CANT_ADD_PARENT_AS_CHILD',
  MASTER_CHILD_BELONGS_TO_OTHER_PARENT = 'MASTER_CHILD_BELONGS_TO_OTHER_PARENT',
  MULTI_LEVEL_ACCOUNT_CHILD_REMOVAL_NOT_SUPPORTED = 'MULTI_LEVEL_ACCOUNT_CHILD_REMOVAL_NOT_SUPPORTED',
  PROFILE_REGISTRATION_SUCCESSFUL_ADD_CHILD_FAILED = 'PROFILE_REGISTRATION_SUCCESSFUL_ADD_CHILD_FAILED',
  PROFILE_NOT_FOUND = 'PROFILE_NOT_FOUND',
  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
  EMAIL_UPDATE_SUCCESSFUL_CART_UPDATE_FAILED = 'EMAIL_UPDATE_SUCCESSFUL_CART_UPDATE_FAILED',
  EMAIL_CHANGE_FAILED = 'EMAIL_CHANGE_FAILED',
}